move Kotlin to own dir
