+++ packages/quicktype-core/src/language/Kotlin/constants.ts
@@ -1,0 +1,51 @@
+export const keywords = [
+    "package",
+    "as",
+    "typealias",
+    "class",
+    "this",
+    "super",
+    "val",
+    "var",
+    "fun",
+    "for",
+    "null",
+    "true",
+    "false",
+    "is",
+    "in",
+    "throw",
+    "return",
+    "break",
+    "continue",
+    "object",
+    "if",
+    "try",
+    "else",
+    "while",
+    "do",
+    "when",
+    "interface",
+    "typeof",
+    "klaxon",
+    "toJson",
+    "Any",
+    "Boolean",
+    "Double",
+    "Float",
+    "Long",
+    "Int",
+    "Short",
+    "System",
+    "Byte",
+    "String",
+    "Array",
+    "List",
+    "Map",
+    "Enum",
+    "Class",
+    "JsonObject",
+    "JsonValue",
+    "Converter",
+    "Klaxon"
+] as const;
