+++ packages/quicktype-core/src/language/Kotlin/index.ts
@@ -1,0 +1,5 @@
+export { KotlinTargetLanguage, kotlinOptions } from "./language";
+export { KotlinRenderer } from "./KotlinRenderer";
+export { KotlinJacksonRenderer } from "./KotlinJacksonRenderer";
+export { KotlinKlaxonRenderer } from "./KotlinKlaxonRenderer";
+export { KotlinXRenderer } from "./KotlinXRenderer";
