+++ packages/quicktype-core/src/language/Kotlin/utils.ts
@@ -1,0 +1,54 @@
+import {
+    allLowerWordStyle,
+    allUpperWordStyle,
+    combineWords,
+    escapeNonPrintableMapper,
+    firstUpperWordStyle,
+    intToHex,
+    isDigit,
+    isLetterOrUnderscore,
+    isNumeric,
+    isPrintable,
+    legalizeCharacters,
+    splitIntoWords,
+    utf32ConcatMap
+} from "../../support/Strings";
+
+function isPartCharacter(codePoint: number): boolean {
+    return isLetterOrUnderscore(codePoint) || isNumeric(codePoint);
+}
+
+function isStartCharacter(codePoint: number): boolean {
+    return isPartCharacter(codePoint) && !isDigit(codePoint);
+}
+
+const legalizeName = legalizeCharacters(isPartCharacter);
+
+export function kotlinNameStyle(
+    isUpper: boolean,
+    original: string,
+    acronymsStyle: (s: string) => string = allUpperWordStyle
+): string {
+    const words = splitIntoWords(original);
+    return combineWords(
+        words,
+        legalizeName,
+        isUpper ? firstUpperWordStyle : allLowerWordStyle,
+        firstUpperWordStyle,
+        isUpper ? allUpperWordStyle : allLowerWordStyle,
+        acronymsStyle,
+        "",
+        isStartCharacter
+    );
+}
+
+function unicodeEscape(codePoint: number): string {
+    return "\\u" + intToHex(codePoint, 4);
+}
+
+const _stringEscape = utf32ConcatMap(escapeNonPrintableMapper(isPrintable, unicodeEscape));
+
+export function stringEscape(s: string): string {
+    // "$this" is a template string in Kotlin so we have to escape $
+    return _stringEscape(s).replace(/\$/g, "\\$");
+}
